MAINT: Refactor tests a bit
