+++ numpy/f2py/tests/test_modules.py
@@ -1,0 +1,35 @@
+import pytest
+import textwrap
+
+from . import util
+from numpy.testing import IS_PYPY
+
+
+@pytest.mark.slow
+class TestModuleDocString(util.F2PyTest):
+    sources = [util.getpath("tests", "src", "modules", "module_data_docstring.f90")]
+
+    @pytest.mark.xfail(IS_PYPY, reason="PyPy cannot modify tp_doc after PyType_Ready")
+    def test_module_docstring(self):
+        assert self.module.mod.__doc__ == textwrap.dedent(
+            """\
+                     i : 'i'-scalar
+                     x : 'i'-array(4)
+                     a : 'f'-array(2,3)
+                     b : 'f'-array(-1,-1), not allocated\x00
+                     foo()\n
+                     Wrapper for ``foo``.\n\n"""
+        )
+
+
+@pytest.mark.slow
+class TestModuleAndSubroutine(util.F2PyTest):
+    module_name = "example"
+    sources = [
+        util.getpath("tests", "src", "modules", "gh25337", "data.f90"),
+        util.getpath("tests", "src", "modules", "gh25337", "use_data.f90"),
+    ]
+
+    def test_gh25337(self):
+        self.module.data.set_shift(3)
+        assert "data" in dir(self.module)
