+++ numpy/f2py/tests/test_regression.py
@@ -66,17 +66,6 @@ def test_include_path():
         assert fname in fnames_in_dir
 
 
-class TestModuleAndSubroutine(util.F2PyTest):
-    module_name = "example"
-    sources = [util.getpath("tests", "src", "regression", "gh25337", "data.f90"),
-               util.getpath("tests", "src", "regression", "gh25337", "use_data.f90")]
-
-    @pytest.mark.slow
-    def test_gh25337(self):
-        self.module.data.set_shift(3)
-        assert "data" in dir(self.module)
-
-
 class TestIncludeFiles(util.F2PyTest):
     sources = [util.getpath("tests", "src", "regression", "incfile.f90")]
     options = [f"-I{util.getpath('tests', 'src', 'regression')}",
